Fix bug

+++ src/main/java/com/teamobi/mobiarmy2/server/ServerManager.java
@@ -21,7 +21,7 @@ public class ServerManager {
 
     private ServerSocket server;
 
-    private int countClients;
+    private long countClients;
     private boolean isStart;
 
     private final ArrayList<ISession> users = new ArrayList<>();
